# openapi.yml
...
openapi: 3.0.3
info:
  title: sombreroquark API
  version: 1.0.0-SNAPSHOT
...
paths:
  /api/auth/login:
    post:
      tags:
      - auth
      requestBody:
        description: User credentials
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                newPassword:
                  type: string
              example:
                email: john.doe@example.com
                password: mysecretpassword
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
              example:
                accessToken: token
                tokenType: Access token
        "400":
          description: Invalid user or password
          content:
            application/json:
              schema:
                type: object
              example:
                error: Invalid user or password
                code: "400"
  /api/auth/login/magic:
    get:
      tags:
      - auth
      parameters:
      - name: redirect
        in: query
        schema:
          type: string
      - name: token
        in: query
        schema:
          type: string
      responses:
        "302":
          description: Redirect to redirect URL
        "400":
          description: Invalid redirect URL
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid redirect URL
                  code: "400"
    post:
      tags:
      - auth
      parameters:
      - name: redirect
        in: query
        description: Redirect URL after login
        schema:
          type: string
        example: https://example.com/dashboard
      requestBody:
        description: Credentials for magic link login
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                newPassword:
                  type: string
              example:
                email: johndoe@example.com
        required: true
      responses:
        "200":
          description: Magic link sent successfully
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Email sent
                  code: "200"
        "400":
          description: Invalid email or redirect URL
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid email or redirect URL
                  code: "400"
  /api/auth/logout:
    post:
      tags:
      - auth
      parameters:
      - name: refreshToken
        in: cookie
        schema:
          type: string
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Logged out
                  code: "200"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid token
                  code: "400"
  /api/auth/token/renew:
    post:
      tags:
      - auth
      parameters:
      - name: refreshToken
        in: cookie
        schema:
          type: string
      responses:
        "200":
          description: Token renewed successfully
          content:
            application/json:
              schema:
                type: object
                example:
                  accessToken: token
                  tokenType: Access token
        "400":
          description: Invalid user or password
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid token
                  code: "400"
...
  /oauth/v1/authorize:
    get:
      tags:
      - OAuth
      parameters:
      - name: provider
        in: query
        schema:
          type: string
      - name: redirect
        in: query
        schema:
          type: string
      responses:
        "302":
          description: Redirect to oauth provider
        "400":
          description: Invalid redirect URL
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid redirect URL
                  code: "400"
  /oauth/v1/callback:
    get:
      tags:
      - OAuth
      parameters:
      - name: code
        in: query
        schema:
          type: string
      - name: state
        in: query
        schema:
          type: string
      responses:
        "302":
          description: Redirect to redirect URL
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                type: object
                example:
                  error: Invalid state
                  code: "400"
...
components:
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT